Fixed Save Command
Save command was saving all the batches  and was a small overlook during
testing. Basically we created a temp process list to store the save into
and we ripped the save from the double linked list, then saved the batch
and then appended the save back to the list. Now we are saving just the
current active batch.