Fixed Kill Command
At first, the kill command was properly killing the batch from the list.
However, the active batch that was removed from the list would still
remain as the active batch when you ran the queue command.  This was
fixed by testing to make sure if the argument 0 was the same as the
active batch. If it was we would test to either switch the active batch
to the next or previous batch.